migrate to section writers for sigv4 and bearer token exec context attributes
